requirements.txt
cryptography==44.0.2
pycryptodome==3.21.0